Split up the tables module
This was getting a bit messy, so I've moved the majority of the
individual tables into their own submodules.
